test(erc721toke): add ERC721Token test script, _setRoyaltyToken & up hh-eth deps

+++ contract/contracts/TokenFactory.sol
@@ -37,8 +37,8 @@ contract TokenFactory is ITokenFactory, ERC2771Context, Multicall, AccessControl
     /// @dev Only CREATOR_ROLE holders.
     bytes32 private constant CREATOR_ROLE = keccak256("CREATOR_ROLE");
 
-    /// @dev Max BPS
-    uint256 private constant MAX_BPS = 10_000;
+    /// @dev The max bps of the contract. So, 10_000 == 100 %
+    uint16 private constant MAX_BPS = 10_000;
 
     /// @dev Token Type => address of the token implemetation.
     mapping(TokenType => address) public tokenBeacons;
